Move files into components, update png
